chore: remove proptypes
